Add files via upload
Bullet Indicator Lights indicate when a laser bullet is ready to shoot. Goes in front of main gun component.
Bullet timer light is a timer that shows how long until the next laser bullet is ready.
There is a background for that as well so the transparency doesnt show through on the main gun component.
Main gun component is the main chassis for the gun UI the bullet timer light and bullet timer background go behind this.
Supershot indicator goes in front of main gun component and houses the large orbs and supershot indicator lights.
The menu selector is the little arrow that sits to the left of the menu items.